More messing with Format
I got it to a "good enough" state of formatting. The width will adjust
to the size of the tree (having overly enough space at high levels) and
there is a trailing empty cell at the end of all rows. The white cells
contain the nodes, and are spaces apart so they are horizontally
centered above their children.